requirements.txt
azure-ai-formrecognizer==3.2.0b5
azure-common==1.1.28
azure-core==1.24.1
boto3==1.24.14
botocore==1.27.14
certifi==2022.6.15
charset-normalizer==2.0.12
idna==3.3
isodate==0.6.1
jmespath==1.0.1
msrest==0.7.1
oauthlib==3.2.0
python-dateutil==2.8.2
requests==2.28.0
requests-oauthlib==1.3.1
s3transfer==0.6.0
six==1.16.0
typing-extensions==4.2.0
urllib3==1.26.9
